+++ Tapjoy/TapjoyRewardedVideoCustomEvent.m
@@ -1,5 +1,6 @@
 
 #import "TapjoyRewardedVideoCustomEvent.h"
+#import "TapjoyAdapterConfiguration.h"
 #import <Tapjoy/Tapjoy.h>
 #import <Tapjoy/TJPlacement.h>
 #if __has_include("MoPub.h")
@@ -7,7 +8,6 @@
     #import "MPLogging.h"
     #import "MPRewardedVideoReward.h"
     #import "MoPub.h"
-    #import "MPRewardedVideoCustomEvent+Caching.h"
 #endif
 #import "TapjoyGlobalMediationSettings.h"
 
@@ -58,8 +58,9 @@
         self.isConnecting = YES;
     }
     else {
-        MPLogInfo(@"Tapjoy rewarded video is initialized with empty 'sdkKey'. You must call Tapjoy connect before requesting content.");
-        [self.delegate rewardedVideoDidFailToLoadAdForCustomEvent:self error:nil];
+        NSError *error = [NSError errorWithCode:MOPUBErrorAdapterFailedToLoadAd localizedDescription:@"Tapjoy rewarded video is initialized with empty 'sdkKey'. You must call Tapjoy connect before requesting content."];
+        MPLogAdEvent([MPLogEvent adLoadFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementName);
+        [self.delegate rewardedVideoDidFailToLoadAdForCustomEvent:self error:error];
     }
 }
 
@@ -75,7 +76,7 @@
     self.placementName = info[@"name"];
 
     // Cache network initialization info
-    [self setCachedInitializationParameters:info];
+    [TapjoyAdapterConfiguration updateInitializationParameters:info];
 
     // Adapter is making connect call on behalf of publisher, wait for success before requesting content.
     if (self.isConnecting) {
@@ -94,7 +95,7 @@
 }
 
 - (void)requestPlacementContentWithAdMarkup:(NSString *)adMarkup {
-    if (self.placementName) {
+    if (self.placementName != nil) {
         self.placement = [TJPlacement placementWithName:self.placementName mediationAgent:@"mopub" mediationId:nil delegate:self];
         self.placement.adapterVersion = MP_SDK_VERSION;
         
@@ -108,23 +109,24 @@
             }
         }
 
+        MPLogAdEvent([MPLogEvent adLoadAttemptForAdapter:NSStringFromClass(self.class) dspCreativeId:nil dspName:nil], self.placementName);
         [self.placement requestContent];
     }
     else {
-        MPLogInfo(@"Invalid Tapjoy placement name specified");
-        NSError *error = [NSError errorWithDomain:MoPubRewardedVideoAdsSDKDomain code:MPRewardedVideoAdErrorInvalidCustomEvent userInfo:nil];
+        NSError *error = [NSError errorWithCode:MOPUBErrorAdapterFailedToLoadAd localizedDescription:@"Invalid Tapjoy placement name specified"];
+        MPLogAdEvent([MPLogEvent adLoadFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementName);
         [self.delegate rewardedVideoDidFailToLoadAdForCustomEvent:self error:error];
     }
 }
 
 - (void)presentRewardedVideoFromViewController:(UIViewController *)viewController {
     if ([self hasAdAvailable]) {
-        MPLogInfo(@"Tapjoy rewarded video will be shown");
+        MPLogAdEvent([MPLogEvent adShowAttemptForAdapter:NSStringFromClass(self.class)], self.placementName);
         [self.placement showContentWithViewController:nil];
     }
     else {
-        MPLogInfo(@"Failed to show Tapjoy rewarded video");
         NSError *error = [NSError errorWithDomain:MoPubRewardedVideoAdsSDKDomain code:MPRewardedVideoAdErrorNoAdsAvailable userInfo:nil];
+        MPLogAdEvent([MPLogEvent adShowFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementName);
         [self.delegate rewardedVideoDidFailToPlayForCustomEvent:self error:error];
     }
 }
@@ -154,9 +156,10 @@
 #pragma mark - TJPlacementDelegate methods
 
 - (void)requestDidSucceed:(TJPlacement *)placement {
+    MPLogAdEvent([MPLogEvent adLoadSuccessForAdapter:NSStringFromClass(self.class)], self.placementName);
     if (!placement.isContentAvailable) {
-        MPLogInfo(@"No Tapjoy rewarded videos available");
         NSError *error = [NSError errorWithDomain:MoPubRewardedVideoAdsSDKDomain code:MPRewardedVideoAdErrorNoAdsAvailable userInfo:nil];
+        MPLogAdEvent([MPLogEvent adShowFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementName);
         [self.delegate rewardedVideoDidFailToLoadAdForCustomEvent:self error:error];
     }
 }
@@ -167,28 +170,30 @@
 }
 
 - (void)requestDidFail:(TJPlacement *)placement error:(NSError *)error {
-    MPLogInfo(@"Tapjoy rewarded video request failed");
+    MPLogAdEvent([MPLogEvent adLoadFailedForAdapter:NSStringFromClass(self.class) error:error], self.placementName);
     [self.delegate rewardedVideoDidFailToLoadAdForCustomEvent:self error:error];
 }
 
 - (void)contentDidAppear:(TJPlacement *)placement {
-    MPLogInfo(@"Tapjoy rewarded video content did appear");
     [Tapjoy setVideoAdDelegate:self];
+    MPLogAdEvent([MPLogEvent adWillAppearForAdapter:NSStringFromClass(self.class)], self.placementName);
     [self.delegate rewardedVideoWillAppearForCustomEvent:self];
+    MPLogAdEvent([MPLogEvent adShowSuccessForAdapter:NSStringFromClass(self.class)], self.placementName);
+    MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], self.placementName);
     [self.delegate rewardedVideoDidAppearForCustomEvent:self];
 }
 
 - (void)contentDidDisappear:(TJPlacement *)placement {
-    MPLogInfo(@"Tapjoy rewarded video content did disappear");
     [Tapjoy setVideoAdDelegate:nil];
+    MPLogAdEvent([MPLogEvent adWillDisappearForAdapter:NSStringFromClass(self.class)], self.placementName);
     [self.delegate rewardedVideoWillDisappearForCustomEvent:self];
+    MPLogAdEvent([MPLogEvent adDidDisappearForAdapter:NSStringFromClass(self.class)], self.placementName);
     [self.delegate rewardedVideoDidDisappearForCustomEvent:self];
 }
 
 #pragma mark Tapjoy Video
 
 - (void)videoAdCompleted {
-    MPLogInfo(@"Tapjoy rewarded video completed");
     [self.delegate rewardedVideoShouldRewardUserForCustomEvent:self reward:[[MPRewardedVideoReward alloc] initWithCurrencyAmount:@(kMPRewardedVideoRewardCurrencyAmountUnspecified)]];
 }